labs/05-analyze-images/python/image-analysis/requirements.txt
python-dotenv
azure-core
debugpy
azure-ai-vision-imageanalysis
pillow
matplotlib